requirements.txt
Flask==2.3.3
replicate==0.10.0
python-dotenv==1.0.0
Pillow==10.0.0
numpy==1.26.0
opencv-python-headless==4.8.0.76
scikit-learn==1.3.0
requests==2.31.0




